refactor: Split types out of ffi.ts and improve readability
